🐼 ajout du bouton close de la popin en savoir plus
- ajout du bouton close sur la popin de confirmation de connection
- update des tests unitaires
- fix du logo de la popin de connexion
