+++ components/connexion/connexion-form.jsx
@@ -5,9 +5,9 @@ import { Fragment, PureComponent } from "react";
 import { Field, Form as FinalForm } from "react-final-form";
 
 import { roles } from "./config.json";
-import EmailInput from "./email-input";
 import ErrorSnackbar from "./error-snackbar";
-import RolesInput from "./roles-input";
+import EmailInput from "./form/email-input";
+import RolesInput from "./form/roles-input";
 import SubmitButton from "./submit-button";
 import { getDefaultRoleFromConfig, updateDomainsWhenRoleChange } from "./utils";
