+++ components/connexion/form/mentions-legales-text.jsx
@@ -9,23 +9,23 @@ const styles = theme => ({
     textDecoration: "underline",
   },
   mentionsLegales: {
+    alignSelf: "flex-end",
     color: "#B1B1B1",
     flex: "1",
     fontFamily: "lato",
     fontSize: "14px",
-    marginBottom: "43px",
-    marginTop: "43px",
-    alignSelf: "flex-end",
-    maxWidth: "52Opx",
-    textAlign: "right",
     letterSpacing: "0.3px",
     lineHeight: "17px",
+    marginBottom: "43px",
+    marginTop: "43px",
+    maxWidth: "52Opx",
+    textAlign: "right",
     [theme.breakpoints.down("xs")]: {
       fontSize: "12px",
+      lineHeight: "14px",
       marginBottom: "20px",
       marginTop: "20px",
       maxWidth: "200px",
-      lineHeight: "14px",
       textAlign: "right",
     },
   },
