--- components/en-savoir-plus/tests/index.test.jsx
@@ -1,20 +1,0 @@
-import { createShallow } from "@material-ui/core/test-utils";
-
-import MentionsLegales from "../index";
-
-describe("components | app-header | index", () => {
-  let shallow;
-  beforeAll(() => {
-    const options = { dive: true };
-    shallow = createShallow(options);
-  });
-
-  describe("snapshot", () => {
-    it("doit correspondre quand l'user n'est pas connecté", () => {
-      const props = {};
-      const wrapper = shallow(<MentionsLegales {...props} />);
-      expect(wrapper).toBeDefined();
-      expect(wrapper).toMatchSnapshot();
-    });
-  });
-});
